requirements.txt
django==5.2.9
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
django-celery-beat==2.8.1
django-celery-results==2.6.0
django-filter==25.2


minio==7.2.19
django-minio-backend==4.4.4

drf-spectacular==0.29.0

psycopg2-binary==2.9.11

redis==7.1.0
celery==5.6.0
gunicorn==23.0.0